requirements_crimson.txt
accelerate==1.13.0
numpy==2.4.4
openai==2.30.0
openpyxl==3.1.5
pandas==3.0.1
peft==0.18.1
Pillow==12.1.1
pydantic==2.12.5
pyyaml==6.0.3
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
torch==2.11.0
tqdm==4.67.3
transformers==5.4.0
